Implement crud operations and move models to models project

+++ Src/Services/ForumService.cs
@@ -1,6 +1,9 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
-using Models;
+using AutoMapper;
+using Models.Business;
+using Models.Request;
 using Repositories;
 
 namespace Services
@@ -8,14 +11,23 @@ namespace Services
     public class ForumService : IForumService
     {
         private readonly IForumRepository _forumRepository;
+        private readonly IMapper _mapper;
 
-        public ForumService(IForumRepository forumRepository)
+        public ForumService(IForumRepository forumRepository, IMapper mapper)
         {
             _forumRepository = forumRepository ?? throw new ArgumentNullException(nameof(forumRepository));
+            _mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
         }
 
-        public async Task<Forum> Create(Forum forum)
+        public Task<List<Forum>> List()
         {
+            return _forumRepository.List();
+        }
+
+        public async Task<Forum> Create(CreateForum create)
+        {
+            var forum = _mapper.Map<Forum>(create);
+
             await _forumRepository.Create(forum);
             return forum;
         }
@@ -25,15 +37,22 @@ namespace Services
             return _forumRepository.Read(id);
         }
 
-        public async Task<Forum> Update(Forum forum)
+        public async Task<Forum> Update(int id, UpdateForum update)
         {
+            var forum = await _forumRepository.Read(id);
+            _mapper.Map(update, forum);
+
             await _forumRepository.Update(forum);
             return forum;
         }
 
-        public Task Delete(Forum forum)
+        public async Task Delete(int id)
         {
-            return _forumRepository.Delete(forum);
+            var forum = await _forumRepository.Read(id);
+            if (forum != null)
+            {
+                await _forumRepository.Delete(forum);
+            }
         }
     }
 }
